FIX: Changed unsigned ints to size_t's

+++ 0x1D-search_algorithms/1-binary.c
@@ -14,7 +14,7 @@
 
 int binary_search(int *array, size_t size, int value)
 {
-	unsigned int i, j,  m = 0, l = 0, r = size - 1;
+	size_t i, j,  m = 0, l = 0, r = size - 1;
 
 	if (array)
 	{
